+++ api/types/plugin/user.go
@@ -1,0 +1,14 @@
+// Code generated from OpenAPI definition. DO NOT EDIT.
+
+package plugin
+
+// User
+type User struct {
+	//
+	// Example: 1000
+	UID uint32 `json:"UID,omitempty"`
+
+	//
+	// Example: 1000
+	GID uint32 `json:"GID,omitempty"`
+}