+++ api/types/storage/root_fs_storage_snapshot.go
@@ -1,0 +1,9 @@
+// Code generated from OpenAPI definition. DO NOT EDIT.
+
+package storage
+
+// RootFSStorageSnapshot Information about a snapshot backend of the container's root filesystem.
+type RootFSStorageSnapshot struct {
+	// Name of the snapshotter.
+	Name string `json:"Name,omitempty"`
+}